requirements.txt
python-dotenv
pillow-heif
Pillow
roboflow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client